Move proof codecs to proof modules, op-proof dispatch to ops/encoding.
